requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
aiohttp==3.9.3
redis==5.0.1
numpy==1.26.4
pydantic==2.6.1
asyncpg==0.29.0
jsonschema==4.20.0
python-dotenv==1.0.0
rank-bm25==0.2.2
scikit-learn==1.4.0